style: rename folders and file name to snake case
